requirements.txt
gspread==5.0.0
oauth2client==4.1.3
mysql-connector-python==8.0.29
pandas==1.5.0
watchdog==2.1.9
